pyproject.toml
[tool.poetry]
name = "vuanem-bq-to-pg"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
google-cloud-bigquery = "^2.30.1"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.11b1"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:warnings -s --durations=0"
testpaths = [
    "test",
]
